requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
openai>=1.54.0
python-dotenv==1.0.0
Werkzeug==3.0.1
gunicorn==22.0.0
requests==2.31.0
httpx>=0.27.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
pytz==2023.3
psycopg2-binary==2.9.9
sendgrid==6.11.0
stripe==7.0.0
